requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15
streamlit>=1.28
pyyaml>=6.0
joblib>=1.3
unidecode>=1.3
scikit-learn>=1.3
jupyter>=1.0
nbformat>=5.9
